moved subcomponents into parent component folder
