+++ services/frontend/service/src/components/MessageInThread/Avatar/index.js
