+++ services/frontend/service/src/components/MessageInThread/Avatar/initialsFromName.js
