+++ services/frontend/service/src/components/MessageInThread/Message/index.js
